post

+++ app/Http/Controllers/TasksController.php
@@ -26,7 +26,7 @@ class TasksController extends Controller
                 'user' => $user,
                 'tasks' => $tasks,
             ];
-            $data += $this->counts($user);
+           /** $data += $this->counts($user); **/
             return view('tasks.index', $data);
         }else {
             return view('welcome');
@@ -56,14 +56,14 @@ class TasksController extends Controller
      */
     public function store(Request $request)
     {
-        $this->validate($request, [
-            'status' => 'required|max:10', 
+         $this->validate($request, [
             'content' => 'required|max:191',
         ]);
-        $task = new Task;
-        $task->status = $request->status; 
-        $task->content = $request->content;
-        $task->save();
+
+        $request->user()->tasks()->create([
+            'content' => $request->content,
+            'status' => $request->status,
+        ]);
 
         return redirect('/');
     }
@@ -76,11 +76,14 @@ class TasksController extends Controller
      */
     public function show($id)
     {
-    $task = Task::find($id);
+     $task = \App\Task::find($id);
 
-        return view('tasks.show', [
-            'task' => $task,
-        ]);
+        if (\Auth::id() === $task->user_id) {
+            return view('tasks.show', [
+            'task' => $task,]);}
+        else{
+            return redirect('/');
+        }
     }
 
     /**
@@ -91,11 +94,14 @@ class TasksController extends Controller
      */
     public function edit($id)
     {
-        $task = Task::find($id);
+       $task = \App\Task::find($id);
 
-        return view('tasks.edit', [
-            'task' => $task,
-        ]);
+        if (\Auth::id() === $task->user_id) {
+            return view('tasks.edit', [
+            'task' => $task,]);}
+        else{
+            return redirect('/');
+        }
     }
 
     /**
@@ -128,9 +134,11 @@ class TasksController extends Controller
      */
     public function destroy($id)
     {
-         $task = Task::find($id);
-        $task->delete();
+        $task = \App\Task::find($id);
 
-        return redirect('/');
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+        }
+         return redirect('/');
     }
 }
